fix: update ignoreWalk to include default ignore rules for common package manager cache directories
This PR updates `ignoreWalk` to include default ignore rules for common package manager cache directories (e.g., `node_modules`).
It introduces `getWorkspaceExcludePatterns` in `packages/common/src/tool-utils/workspace-exclude-patterns.ts` which provides a comprehensive list of patterns to exclude.
This function is then used in:
- `packages/vscode/src/integrations/checkpoint/shadow-git-excludes.ts`
- `packages/vscode/src/integrations/webview/vscode-host-impl.ts`
- `packages/cli/src/lib/read-environment.ts`

This ensures that `node_modules` and other cache directories are ignored by default when listing files in the workspace, improving performance and reducing noise.
