+++ packages/common/src/tool-utils/index.ts
@@ -40,3 +40,4 @@ export {
 } from "./notebook-utils";
 export { readMediaFile } from "./media";
 export { resolveToolCallArgs, resolvePochiUri } from "./resolve-tool-call-args";
+export { getWorkspaceExcludePatterns } from "./workspace-exclude-patterns";
